% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim2fit.R
\name{sim2fit}
\alias{sim2fit}
\title{Simulate TVVAR model and add observation error}
\usage{
sim2fit(obj, n_sims, sd = 0.1)
}
\arguments{
\item{obj}{A fitted \code{simTVVAR} object.}

\item{n_sims}{The number of realizations of the TVVAR process.}

\item{sd}{The standard deviation of the Gaussian observation errors.
Can be set to 0 for no observation error.}
}
\value{
An array with dimensions \code{c(n_sim, TT, n_spp)}.
}
\description{
\code{sim2fit} adds observation error to a simulated TVVAR process and
  converts it to a form suitable for fitting with \code{tvvarss}.
}
\details{
This is a helper function that takes a fitted \code{simTVVAR} object and
simulates multiple realizations of the process before adding Gaussian
obsveration errors.
}
\examples{
set.seed(123)
## number of time steps
TT <- 30
## number of spp/guilds
nn <- 4
## CASE 1: linear food chain
B0_lfc <- matrix(list(0),nn,nn)
for(i in 1:(nn-1)) {
  B0_lfc[i,i+1] <- "td"
  B0_lfc[i+1,i] <- "bu"
}
## simulate process
lfc <- simTVVAR(B0_lfc, TT, var_QX=rev(seq(1,4)/40), cov_QX=0, var_QB=0.05, cov_QB=0)
## create data array with 3 realizations of the process
dat <- sim2fit(lfc, 3)

}
