\name{tvvarss-package}
\alias{tvvarss-package}
\alias{tvvarss}
\docType{package}
\title{
\packageTitle{tvvarss}
}
\description{
\packageDescription{tvvarss}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{tvvarss}
\packageIndices{tvvarss}
}
\author{
\packageAuthor{tvvarss}

Maintainer: \packageMaintainer{tvvarss}
}
\references{

Francis, T.B. E. Wolkovich, S.E. Hampton, M.D. Scheuerell, S.L. Katz, and E.E. Holmes. 2014. \emph{Shifting regimes and changing interactions in the Lake Washington, USA, plankton community from 1962–1994}. PLoS ONE 9(10): e110363. doi:10.1371/journal.pone.0110363.

Hampton, S. E., Holmes, E. E., Scheef, L. P., Scheuerell, M. D., Katz, S. L., Pendleton, D. E. and Ward, E. J. 2013. \emph{Quantifying effects of abiotic and biotic drivers on community dynamics with multivariate autoregressive (MAR) models}. Ecology, 94: 2663–2669. doi:10.1890/13-0996.1

Holmes, E. E., E. J. Ward, and M. D. Scheuerell. 2012. \emph{Analysis of multivariate time-series using the MARSS package}. NOAA Fisheries, Northwest Fisheries Science Center, 2725 Montlake Blvd E., Seattle, WA 98112

Holmes, E. E., E. J. Ward and K. Wills. 2012. \emph{MARSS: Multivariate autoregressive state-space models for analyzing time-series data}. R Journal 4: 11-19.

Ives, A.R., B. Dennis, K.L. Cottingham, and S.R. Carpenter. 2003. \emph{Estimating community stability and ecological interactions from time-series data}. Ecological Monographs, 73(2), pp. 301–330.

Ives, A.R. and V. Dakos. 2012. \emph{Detecting dynamical changes in nonlinear time series using locally linear state-space models}. Ecosphere 3(6):58. http://dx.doi.org/10.1890/ES11-00347.1
}
\keyword{ package }
\seealso{
Optional links to other man pages, e.g.
}
\examples{
#simple examples of the most important functions
}
