% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTVVAR.R
\name{simTVVAR}
\alias{simTVVAR}
\title{Simulate the process component of a TVVARSS model}
\usage{
simTVVAR(
  Bt = NULL,
  topo = NULL,
  TT,
  var_QX,
  cov_QX,
  var_QB,
  cov_QB = 0,
  QQ_XX = NULL,
  QQ_BB = NULL,
  X0 = NULL,
  CC = NULL,
  cc = NULL
)
}
\arguments{
\item{Bt}{A matrix describing the topology of the food web (see 'Details').
If \code{Bt == NULL}, then the food web topology must be specified and
passed as \code{topo}. See 'Details'.}

\item{topo}{Optional list matrix describing the presumed topology of the
community. Pairwise interactions are specified as density-dependent ("dd"),
top-down ("td"), bottom-up ("bu"), competitive/facilitative ("cf"), or
absent ("zero"). If specified, pairwise interactions will be constrained
in an approporiate manner (e.g., top-down effects are between -1 and 0).}

\item{TT}{Number of time steps to simulate.}

\item{var_QX}{Scalar or vector of variances for process errors of states.}

\item{cov_QX}{Covariance, if any, of the process errors of the states; if \code{cov_QX} > 0, then \code{var_QX} must be a scalar.}

\item{var_QB}{Scalar or vector of variances for process errors of \strong{B}.}

\item{cov_QB}{Covariance, if any, of process errors of \strong{B}; if \code{cov_QB} > 0, then \code{var_QB} must be a scalar.}

\item{QQ_XX}{Optionally specify the explicit form for the var-cov matrix \strong{Q} of the process errors of the states.}

\item{QQ_BB}{Optionally specify the explicit form for the var-cov matrix \strong{Q} of the process errors of \strong{B}.}

\item{X0}{Optionally specify vector of initial states; \code{nrow(X0)} must equal \code{nrow(Bt)}.}

\item{CC}{Optionally specify matrix of covariate effects on states.}

\item{cc}{Optionally specify matrix of covariates.}
}
\value{
A list with the following components:
\describe{
\item{\code{B_mat}}{An array of the \strong{B} matrix over time; \code{dim(B_mat) = c(n,n,T+1)}.}
\item{\code{WW_BB}}{The process errors for \strong{B}; \code{dim(WW_BB) = c(n^2,T)}.}
\item{\code{QQ_BB}}{Variance-covariance matrix of the process errors for \strong{B}; \code{dim(QQ_BB) = c(n^2,n^2)}.}
\item{\code{states}}{A matrix of the states over time; \code{dim(states) = c(n,T+1)}.}
\item{\code{WW_XX}}{The process errors (innovations) for the states; \code{dim(WW_XX) = c(n,T)}.}
\item{\code{QQ_XX}}{Variance-covariance matrix of the process errors for the states; \code{dim(QQ_XX) = c(n,n)}.}
\item{\code{call}}{The function call as returned by \code{match.call()}.}
}
}
\description{
\code{simTVVAR} simulates the process (state) component of a TVVARSS model.
}
\details{
\code{Bt} can be used in one of two ways when simulating a TVVAR model:
\enumerate{
\item An \eqn{n x n} \code{matrix} with initial numeric values of B (i.e., B0).
If \code{QQ_BB = matrix(0, n, n)} then, a time-invariant (MARSS) model is
simulated based on these values.
\item An \eqn{n x n x (T+1)} \code{array} with actual values of B for each time
step, including B0. This is useful for simulating multiple realizations
of the same process.
}
\code{topo} can be used to specify the food web topology by passing an
\eqn{n x n} \code{matrix} with a combination of \code{character} and
\code{numeric} values in the off-diagonal elements; the diagonal should
always contain \code{"dd"} as density-dependence is implicit in this
model. Use 0 or "zero" to indicate no interaction and the following
\code{character} codes for ecological interactions:
\itemize{
\item \code{"td"} to indicate a top-down interaction
\item \code{"bu"} to indicate a bottom-up interaction
\item \code{"cf"} to indicate a competitive/facilitative
interaction
}
See 'Examples' for details on formatting \code{B0}.
}
\examples{

}
