% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvvarss.R
\name{tvvarss}
\alias{tvvarss}
\title{tvvarss is the primary function for fitting tvvarss models to time series data.}
\usage{
tvvarss(y, include_trend = TRUE, de_mean = TRUE, B = NULL, x0 = NULL,
  shared_q = NULL, shared_r = NULL, shared_u = NULL, mcmc_iter = 1000,
  mcmc_warmup = 500, mcmc_thin = 1, mcmc_chain = 3, family = "gaussian")
}
\arguments{
\item{y}{The data (array, with dimensions = site, year, species)}

\item{include_trend}{Whether to include time trends. Defaults to TRUE}

\item{de_mean}{Whether or not to de_mean the process model, defaults to TRUE. For example, X_t+1 = B * (X_t - pred[X_t]) versus X_t+1 = B * (X_t)}

\item{B}{The list matrix describing optionally whether elements are 'zero', top-down ('td'), bottom up ('bu'), competitive-facilitative ('cf')}

\item{x0}{The location matrix (mean) of priors on initial states. Defaults to centered on observed data}

\item{shared_q}{Optional matrix (number of species x number of sites) with integers indicating which process variance parameters are shared. Defaults to shared process variances across sites, but unique to each species}

\item{shared_r}{Optional matrix (number of species x number of sites) with integers indicating which observation variance parameters are shared. Defaults to shared observation variances across sites, but unique to each species}

\item{shared_u}{Optional matrix (number of species x number of sites) with integers indicating which trend parameters are shared. Defaults to shared trends across sites, but unique to each species}

\item{mcmc_iter}{Number of MCMC iterations, defaults to 1000}

\item{mcmc_warmup}{Warmup / burn in phase, defaults to 500}

\item{mcmc_thin}{MCMC thin, defaults to 1}

\item{mcmc_chain}{MCMC chains, defaults to 3}

\item{family}{observation model, defaults to 'gaussian'}
}
\value{
an object of class 'stanfit'
}
\description{
tvvarss is the primary function for fitting tvvarss models to time series data.
}

